# Replace np.searchsorted with bisect in indexing