Increases indexing performance, and is consistent with general `bisect` usage in the rest of the file.

By using `bisect` instead of `np.searchsorted`, v-indexing performance increases significantly. Since the search index (`ind`) seems to be a always scalar, `bisect` is a drop-in replacement.

Test case:
```
import dask.array as da
idx = np.random.randint(0, 10000, size=1000)
x = da.random.random((10000,), chunks=(10,))

%timeit x.vindex[idx]
```
Before: `57 ms ± 337 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)`

After: `4.17 ms ± 217 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`

- [x] Tests ~~added~~ / passed
- [x] Passes ~~`black dask`~~ / `flake8 dask`

(I do see one test failure, but it happens also without the change).