@jakirkham 2020-10-05
Thanks Joachim! This is a nice improvement. I wonder if there are [other places we could use this strategy]( https://github.com/dask/dask/search?q=searchsorted ).