@jobh 2020-10-05
Yeah, in superficial testing the upside of `searchsorted` is mainly when the number of search indices is large - with a single "needle" (search index) and a small haystack, `bisect` is 50x performance; with a single needle and haystack of 1M, `searchsorted` is 2x performance; with 1M needles into a 1M haystack, `searchsorted` is 50x performance.